extract `bikeroutes` processing from `roads_bikelanes`
